requirements.txt
pillow==11.0.0
django-filter==24.3
Django==4.2.15
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
daphne==4.0.0
psycopg2-binary
drf-yasg==1.21.7
google-auth==2.26.1
requests==2.31.0
pandas==2.2.3
phonenumbers==8.13.43
django-admin-interface==0.28.5
openpyxl==3.1.5
whitenoise==6.7.0
gunicorn==23.0.0
gevent==24.2.1